Langchain_Crash_Course/requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
openai==1.48.0
langchain-core
langchain_google_firestore
firebase-admin
python-dotenv
chromadb
langchain_chroma
bs4
tavily-python